pyproject.toml
[project]
name = "clyro"
version = "0.2.0"
description = "Runtime governance for AI agents — prevent failures before they happen"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Clyro Team", email = "engineering@clyrohq.com" }
]
keywords = [
    "ai", "agents", "governance", "observability", "guardrails",
    "langchain", "langgraph", "crewai", "anthropic", "claude",
    "mcp", "model-context-protocol", "cost-tracking", "policy",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    # HTTP Client
    "httpx>=0.27.0",
    # Data Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Logging
    "structlog>=24.0.0",
    # Retry Logic
    "tenacity>=8.0.0",
    # Local Storage (SQLite for offline traces)
    "aiosqlite>=0.19.0",
    # YAML config loading (MCP wrapper + hooks)
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Token estimation for cost calculation
tiktoken = [
    "tiktoken>=0.5.0",
]
# OTLP export support
otlp = [
    "opentelemetry-proto>=1.20.0",
]
# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-watch>=4.2.0",
    "ruff>=0.4.0",
    "tiktoken>=0.5.0",
    "opentelemetry-proto>=1.20.0",
]

[project.scripts]
clyro-sdk = "clyro.cli:main"
clyro-mcp = "clyro.mcp.cli:main"
clyro-hook = "clyro.hooks.cli:main"

[project.urls]
Homepage = "https://clyrohq.com"
Documentation = "https://docs.clyrohq.com"
Repository = "https://github.com/getclyro/clyro"
Issues = "https://github.com/getclyro/clyro/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["clyro*"]

[tool.pytest.ini_options]
testpaths = ["tests/sdk", "tests/mcp", "tests/hooks", "tests/integration"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
